refactor: extract cli into cli.go preserving exit-1-on-any-failure
